resources/github-issues/pyproject.toml
[project]
name = "ol-concourse-github-issues"
version = "0.2.0"
description = "Concourse resource type for managing GitHub Issues"
authors = [{ name = "MIT Open Learning", email = "odl-devops@mit.edu" }]
requires-python = ">=3.13"
readme = "README.md"
license = "BSD-3-Clause"
dependencies = [
    "concoursetools>=0.8.0,<0.9",
    "PyGithub>=2.0.0,<3",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0,<9",
    "pytest-mock>=3.14.0,<4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
sources = ["."]
include = ["concourse.py"]
